imagemagick: reintroduce package, cleanup and upgrade
ImageMagick is a free and open-source software suite for displaying,
converting, and editing raster image and vector image files. This
package provides the prominent 'convert' utility.

ImageMagick's installation footprint is:
  x86_64: 6.7 MB,
  ipq806x (ARM): 6.0 MB,
  bcm53xx: 6.2 MB.

The shared libraries occupy 2.4 MB on bcm53xx, 2.3 MB on ipq806x, and
2.5 MB on x86_64. The 114 ImageMagick's modules occupy from 3.4 MB to
4.0 MB depending on the target. It may be possible to reduce the
installation footprint by introducing build parameters to control the
selection of modules. In view of the large number of modules and the
possibility of breakage due to module interdependencies or other
reasons, such attempt is not made at this time.

ImageMagick is therefore best suited for extroot-enabled or x86_64
OpenWrt systems.

In many cases, GraphicsMagick may be used as a substitute for
ImageMagick. GraphicsMagick provides similar functionality, it is
faster and it has a smaller installation footprint. It is therefore
better suited for non-extroot OpenWrt systems. However, in tests to
reduce resolution (and size) of a high-resolution JPEG image
GraphicsMagick required about 25% more RAM than ImageMagick (no HDRI,
quantum depth of 8) during its execution.

Signed-off-by: Val Kulkov <val.kulkov@gmail.com>