nss: fix compilation with QUILT
Commit 657574f45 disabled building the host nsinstall program when using
QUILT.  However, the host nsinstall is needed to compile the package,
breaking compilation with QUILT.

Move the native compile to Build/Configure, which will not be called for
prepare, refresh, or update targets, but will be called before
Build/Compile.

nss does not have a configure script, so Build/Configure/Default is not
being called.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>